removing cast to float to see if fix worked
git-svn-id: file:///d/Documents/workspace/gemserksvn/var/svn/games@38 1dcdb8b7-3cac-4752-9c2b-54fd448f9028

+++ towerofdefense/trunk/game/src/main/java/com/gemserk/games/towerofdefense/components/GuiLogicComponent.java
@@ -109,8 +109,8 @@ public class GuiLogicComponent extends ReflectionComponent {
 		Vector2f mousePosition = mousePositionProperty.getValue(entity);
 
 		if (message.getId().equals("move")) {
-			float x = (Float) Properties.getValue(message, "x");
-			float y = (Float) Properties.getValue(message, "y");
+			float x = Properties.getValue(message, "x");
+			float y = Properties.getValue(message, "y");
 			mousePosition.set(x, y);
 		}
